Cleaned up model - static members and methods control it now
